feat(cmd/unhealthy-tgs): Tidy up source code
